Use u32 for inverted index doc ids

+++ search/src/FinetunableRetriever.h
@@ -5,7 +5,7 @@
 
 namespace thirdai::search {
 
-using QueryId = uint64_t;
+using QueryId = uint32_t;
 
 class FinetunableRetriever {
  public:
